+++ codecin/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "5.5.0"
+__version__ = "5.5.3"
 __author__ = "ByUsi Studio"
 
 from .config import Config